Create a NAT gateway per availability zone
This changes the way that utility subnets work. Instead of a single
utility subnet, there is now there is a utility subnet per AZ. A NAT
is deployed into each of these utility subnets. A private route table
is created for each NAT and associated with all private subnets in that
AZ.